Updated the signatures of some parsing functions
It was required to change the type of some string position index variables from int to ssize_t for consistency reasons.
It should be noted that the position index variables for these strings should be signed integers as negative values (-1) are also possible

+++ connectioninputparser.cpp
@@ -141,11 +141,11 @@ void ConnectionInputParser::_reset()
     Parser::_reset();
 }
 
-int ConnectionInputParser::_parseCablePartNumber(const size_t rowIndex, const int currentPosition)
+ssize_t ConnectionInputParser::_parseCablePartNumber(const size_t rowIndex, const ssize_t currentPosition)
 {
     assert(rowIndex < mInputData.size());
 
-    const int resultingPosition{readDataField(mInputData[rowIndex], mCablePartNumbersEntries[rowIndex], currentPosition)};
+    const ssize_t resultingPosition{readDataField(mInputData[rowIndex], mCablePartNumbersEntries[rowIndex], currentPosition)};
 
     // if no cable PN entered on current row take the PN for previous row
     if (0u == mCablePartNumbersEntries[rowIndex].size())
